refactor : 함께 가기 알림 요청 보냈는지 확인하는 메서드를 RequestNotificationCommandService -> RequestNotificationQueryService로 이동
#312
